+++ src/modules/beatport-link.js
@@ -1,0 +1,103 @@
+const PATH = require('path');
+const FIRSTLINE = require('firstline');
+
+var lastID = 0;
+
+module.exports = {
+    monitor: function (path) {
+        fs.readdir(path, function(err, files) {
+            if (err) {
+                return console.log('Unable to scan directory: ' + err);
+            }
+            else { /*console.log ("Scanning: " + rekordbox_path);*/ }
+    
+            files.forEach(function(file) {
+                let result = FIRSTLINE(PATH.join(path, file));
+    
+                result.then(
+                    function(resolve) {
+    
+                        var object = JSON.parse(extractJSON(resolve));
+    
+                        if (object.id == lastID) {
+                            return;
+                        }
+                        else {
+                            lastID = object.id;
+                            console.log(object);
+                            
+                            artist = formatArtist(object);
+                            title = formatTitle(object);
+    
+                            console.log(artist + " - " + title);
+                            document.getElementById('bl_current').innerHTML = artist + " - " + title;
+                        }
+                        
+                    },
+                    function(error) {
+                        console.log("The promise could not be resolved: " + error);
+                    }
+                )
+                return;
+            });
+        });
+    }
+}
+
+function formatTitle(object) {
+    if (object.remixers.length > 0) {
+        title = object.name + ' (' + object.mix_name + ')';
+    }
+    else {
+        title = object.name;
+    }
+}
+
+function formatArtist(object) {
+    let artist;
+    for (let i = 0; i < object.artists.length; i++) {
+    
+        if (i == object.artists.length - 1) {
+            artist += object.artists[i].name;
+            break;
+        }
+        if (i == object.artists.length - 2) {
+            artist += object.artists[i].name + " & ";
+        }
+        else {
+            artist += object.artists[i].name + ", ";
+        }
+    }
+    return artist;
+}
+
+function extractJSON(input) {
+    let array = input.split('');
+    let opening_braces = 0, closing_braces = 0, start_pos = 0, end_pos = 0;
+    
+    for (let i = 0; i < array.length; i++) {
+        if (array[i] == '{') {
+            opening_braces++;
+    
+            if (opening_braces == 1) {
+                start_pos = i;
+            }
+        }
+        if (array[i] == '}') {
+            closing_braces++;
+            if (closing_braces == opening_braces) {
+                end_pos = i;
+                break;
+            }
+        }
+    }
+    
+    if (start_pos == 0 && end_pos == 0) {
+        return;
+    }  
+    
+    array.splice(0, start_pos);
+    array.splice(end_pos, array.length);
+    
+    return resolve.substring(start_pos, end_pos+1);
+}
